+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.43] - 2026.09.14
 
 ### Added
 
@@ -429,6 +429,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Agent TOML configuration example
 
 [Unreleased]: https://github.com/studiometa/trafic/compare/0.1.39...HEAD
+[0.1.43]: https://github.com/studiometa/trafic/compare/0.1.42...0.1.43
 [0.1.42]: https://github.com/studiometa/trafic/compare/0.1.41...0.1.42
 [0.1.41]: https://github.com/studiometa/trafic/compare/0.1.40...0.1.41
 [0.1.40]: https://github.com/studiometa/trafic/compare/0.1.39...0.1.40